recursively search upwards for pnp.cjs
